cool-magazine-workspace/libs/domains/generic/identity/oidc-core/src/lib/oidc-1.0-bounded-context/aggregate-roots/client/validation-rules/secret-property/secret-property-client-registration-data-validation-rule.ts
import { ValidationRuleBase } from '@bh2old/ddd-expc';
import { IClientRegistrationData } from '../../models';
import { SecretPropertyClientRegistrationDataSpecification } from '../../specifications';

export class SecretPropertyClientRegistrationDataValidationRule extends ValidationRuleBase<IClientRegistrationData> {
  constructor() {
    const specification =
      new SecretPropertyClientRegistrationDataSpecification();
    const message =
      `property: 'must be either string value or undefined', ` +
      `string value: 'must be at least ${SecretPropertyClientRegistrationDataSpecification.STRING_LENGTH.min} in length'`;
    const property = 'secret';

    super(specification, message, property);
  }
}
